fixed the following bug:
Our repos are singletons, meaning that LiveData will not change if saved in the class. This is useful for screen rotations since the old data is still retained. However, when e.g. sending ValidationEntries the old LiveData might have the State SUCCESS, triggering a false observe success, therefore for cases like this one needs to recreate the LiveData to prevent old data from interfering